pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lg"
version = "0.1.0"
description = "Linear + Git helper CLI"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv",
  "requests",
  "InquirerPy"
]

[project.scripts]
lg = "main:run"

[tool.setuptools]
py-modules = ["main", "controller"]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*"]